agents/pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anchorbrowser>=0.1.0",
    "browser-use>=0.9.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "convex>=0.5.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.10.0",
]
